+++ frontend/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "axios": "^1.8.4",
     "chart.js": "^4.4.8",
+    "frontend": "file:",
     "react": "^19.0.0",
     "react-chartjs-2": "^5.3.0",
     "react-dom": "^19.0.0",
