+++ iot_backend/config/alert_config.yaml
@@ -1,0 +1,70 @@
+# Alert System Configuration
+
+# Channels Configuration
+channels:
+  email:
+    enabled: true
+    sender: "iot.alerts@yourdomain.com"
+    recipients: ["abhinav@yourdomain.com"]  # Your email
+    smtp_server: "smtp.yourdomain.com"
+    smtp_port: 587
+    use_tls: true
+  webhook:
+    enabled: true
+    url: "http://localhost:8000/api/alerts/webhook"
+  in_app:
+    enabled: true
+    websocket_url: "ws://localhost:8000/ws/alerts"
+
+# Anomaly Detection Thresholds
+thresholds:
+  temperature:
+    high: 30.0
+    low: 10.0
+    anomaly_sensitivity: 0.1  # IsolationForest contamination parameter
+  humidity:
+    high: 80.0
+    low: 20.0
+    anomaly_sensitivity: 0.1
+  pressure:
+    high: 1100.0
+    low: 900.0
+    anomaly_sensitivity: 0.1
+
+# Alert Rules
+alert_rules:
+  - name: "temperature_spike"
+    condition: "value > thresholds.temperature.high"
+    severity: "high"
+    message: "Temperature exceeded threshold: {value}°C"
+    cooldown_period: 300  # 5 minutes between repeated alerts
+    
+  - name: "temperature_anomaly"
+    condition: "is_anomaly"
+    severity: "medium"
+    message: "Unusual temperature pattern detected: {value}°C"
+    cooldown_period: 600  # 10 minutes between anomaly alerts
+    
+  - name: "humidity_extreme"
+    condition: "value > thresholds.humidity.high or value < thresholds.humidity.low"
+    severity: "high"
+    message: "Humidity outside normal range: {value}%"
+    cooldown_period: 300
+
+# Notification Settings
+notification_settings:
+  batch_interval: 60  # Seconds to batch similar alerts
+  max_alerts_per_hour: 50  # Rate limiting for alerts
+  retention_days: 30  # How long to keep alert history
+
+# Alert Priorities
+severity_levels:
+  low:
+    color: "#2196F3"
+    notification_channels: ["in_app"]
+  medium:
+    color: "#FFC107"
+    notification_channels: ["in_app", "email"]
+  high:
+    color: "#F44336"
+    notification_channels: ["in_app", "email", "webhook"]
