+++ iot_backend/timeseries/management/commands/import_timeseries.py
